commander: calibration routines delete unused sphere fit
 - consume remaining matrix alg dependencies
