+++ src/lib/mathlib/CMakeLists.txt
@@ -33,7 +33,6 @@
 
 px4_add_library(mathlib
 	math/test/test.cpp
-	math/matrix_alg.cpp
 	math/filter/LowPassFilter2p.cpp
 	math/filter/LowPassFilter2pVector3f.cpp
 )
