+++ src/lib/mathlib/mathlib.h
@@ -43,7 +43,6 @@
 
 #include "math/Limits.hpp"
 #include "math/Functions.hpp"
-#include "math/matrix_alg.h"
 #include "math/SearchMin.hpp"
 #include "math/TrajMath.hpp"
