+++ src/modules/commander/calibration_routines.cpp
@@ -60,179 +60,215 @@
 
 using namespace time_literals;
 
-int sphere_fit_least_squares(const float x[], const float y[], const float z[],
-			     unsigned int size, unsigned int max_iterations, float delta, float *sphere_x, float *sphere_y, float *sphere_z,
-			     float *sphere_radius)
+
+/*
+ *    Does matrix multiplication of two regular/square matrices
+ *
+ *    @param     A,           Matrix A
+ *    @param     B,           Matrix B
+ *    @param     n,           dimemsion of square matrices
+ *    @returns                multiplied matrix i.e. A*B
+ */
+static float *mat_mul(float *A, float *B, uint8_t n)
 {
+	float *ret = new float[n * n];
+	memset(ret, 0.0f, n * n * sizeof(float));
 
-	float x_sumplain = 0.0f;
-	float x_sumsq = 0.0f;
-	float x_sumcube = 0.0f;
-
-	float y_sumplain = 0.0f;
-	float y_sumsq = 0.0f;
-	float y_sumcube = 0.0f;
-
-	float z_sumplain = 0.0f;
-	float z_sumsq = 0.0f;
-	float z_sumcube = 0.0f;
-
-	float xy_sum = 0.0f;
-	float xz_sum = 0.0f;
-	float yz_sum = 0.0f;
-
-	float x2y_sum = 0.0f;
-	float x2z_sum = 0.0f;
-	float y2x_sum = 0.0f;
-	float y2z_sum = 0.0f;
-	float z2x_sum = 0.0f;
-	float z2y_sum = 0.0f;
-
-	for (unsigned int i = 0; i < size; i++) {
-
-		float x2 = x[i] * x[i];
-		float y2 = y[i] * y[i];
-		float z2 = z[i] * z[i];
-
-		x_sumplain += x[i];
-		x_sumsq += x2;
-		x_sumcube += x2 * x[i];
-
-		y_sumplain += y[i];
-		y_sumsq += y2;
-		y_sumcube += y2 * y[i];
-
-		z_sumplain += z[i];
-		z_sumsq += z2;
-		z_sumcube += z2 * z[i];
-
-		xy_sum += x[i] * y[i];
-		xz_sum += x[i] * z[i];
-		yz_sum += y[i] * z[i];
-
-		x2y_sum += x2 * y[i];
-		x2z_sum += x2 * z[i];
-
-		y2x_sum += y2 * x[i];
-		y2z_sum += y2 * z[i];
-
-		z2x_sum += z2 * x[i];
-		z2y_sum += z2 * y[i];
+	for (uint8_t i = 0; i < n; i++) {
+		for (uint8_t j = 0; j < n; j++) {
+			for (uint8_t k = 0; k < n; k++) {
+				ret[i * n + j] += A[i * n + k] * B[k * n + j];
+			}
+		}
 	}
 
-	//
-	//Least Squares Fit a sphere A,B,C with radius squared Rsq to 3D data
-	//
-	//    P is a structure that has been computed with the data earlier.
-	//    P.npoints is the number of elements; the length of X,Y,Z are identical.
-	//    P's members are logically named.
-	//
-	//    X[n] is the x component of point n
-	//    Y[n] is the y component of point n
-	//    Z[n] is the z component of point n
-	//
-	//    A is the x coordiante of the sphere
-	//    B is the y coordiante of the sphere
-	//    C is the z coordiante of the sphere
-	//    Rsq is the radius squared of the sphere.
-	//
-	//This method should converge; maybe 5-100 iterations or more.
-	//
-	float x_sum = x_sumplain / size;        //sum( X[n] )
-	float x_sum2 = x_sumsq / size;    //sum( X[n]^2 )
-	float x_sum3 = x_sumcube / size;    //sum( X[n]^3 )
-	float y_sum = y_sumplain / size;        //sum( Y[n] )
-	float y_sum2 = y_sumsq / size;    //sum( Y[n]^2 )
-	float y_sum3 = y_sumcube / size;    //sum( Y[n]^3 )
-	float z_sum = z_sumplain / size;        //sum( Z[n] )
-	float z_sum2 = z_sumsq / size;    //sum( Z[n]^2 )
-	float z_sum3 = z_sumcube / size;    //sum( Z[n]^3 )
+	return ret;
+}
 
-	float XY = xy_sum / size;        //sum( X[n] * Y[n] )
-	float XZ = xz_sum / size;        //sum( X[n] * Z[n] )
-	float YZ = yz_sum / size;        //sum( Y[n] * Z[n] )
-	float X2Y = x2y_sum / size;    //sum( X[n]^2 * Y[n] )
-	float X2Z = x2z_sum / size;    //sum( X[n]^2 * Z[n] )
-	float Y2X = y2x_sum / size;    //sum( Y[n]^2 * X[n] )
-	float Y2Z = y2z_sum / size;    //sum( Y[n]^2 * Z[n] )
-	float Z2X = z2x_sum / size;    //sum( Z[n]^2 * X[n] )
-	float Z2Y = z2y_sum / size;    //sum( Z[n]^2 * Y[n] )
+static constexpr void swap(float &a, float &b)
+{
+	float c = a;
+	a = b;
+	b = c;
+}
 
-	//Reduction of multiplications
-	float F0 = x_sum2 + y_sum2 + z_sum2;
-	float F1 =  0.5f * F0;
-	float F2 = -8.0f * (x_sum3 + Y2X + Z2X);
-	float F3 = -8.0f * (X2Y + y_sum3 + Z2Y);
-	float F4 = -8.0f * (X2Z + Y2Z + z_sum3);
-
-	//Set initial conditions:
-	float A = x_sum;
-	float B = y_sum;
-	float C = z_sum;
-
-	//First iteration computation:
-	float A2 = A * A;
-	float B2 = B * B;
-	float C2 = C * C;
-	float QS = A2 + B2 + C2;
-	float QB = -2.0f * (A * x_sum + B * y_sum + C * z_sum);
-
-	//Set initial conditions:
-	float Rsq = F0 + QB + QS;
-
-	//First iteration computation:
-	float Q0 = 0.5f * (QS - Rsq);
-	float Q1 = F1 + Q0;
-	float Q2 = 8.0f * (QS - Rsq + QB + F0);
-	float aA, aB, aC, nA, nB, nC, dA, dB, dC;
-
-	//Iterate N times, ignore stop condition.
-	unsigned int n = 0;
-
-	while (n < max_iterations) {
-		n++;
-
-		//Compute denominator:
-		aA = Q2 + 16.0f * (A2 - 2.0f * A * x_sum + x_sum2);
-		aB = Q2 + 16.0f * (B2 - 2.0f * B * y_sum + y_sum2);
-		aC = Q2 + 16.0f * (C2 - 2.0f * C * z_sum + z_sum2);
-		aA = (fabsf(aA) < FLT_EPSILON) ? 1.0f : aA;
-		aB = (fabsf(aB) < FLT_EPSILON) ? 1.0f : aB;
-		aC = (fabsf(aC) < FLT_EPSILON) ? 1.0f : aC;
-
-		//Compute next iteration
-		nA = A - ((F2 + 16.0f * (B * XY + C * XZ + x_sum * (-A2 - Q0) + A * (x_sum2 + Q1 - C * z_sum - B * y_sum))) / aA);
-		nB = B - ((F3 + 16.0f * (A * XY + C * YZ + y_sum * (-B2 - Q0) + B * (y_sum2 + Q1 - A * x_sum - C * z_sum))) / aB);
-		nC = C - ((F4 + 16.0f * (A * XZ + B * YZ + z_sum * (-C2 - Q0) + C * (z_sum2 + Q1 - A * x_sum - B * y_sum))) / aC);
-
-		//Check for stop condition
-		dA = (nA - A);
-		dB = (nB - B);
-		dC = (nC - C);
-
-		if ((dA * dA + dB * dB + dC * dC) <= delta) { break; }
-
-		//Compute next iteration's values
-		A = nA;
-		B = nB;
-		C = nC;
-		A2 = A * A;
-		B2 = B * B;
-		C2 = C * C;
-		QS = A2 + B2 + C2;
-		QB = -2.0f * (A * x_sum + B * y_sum + C * z_sum);
-		Rsq = F0 + QB + QS;
-		Q0 = 0.5f * (QS - Rsq);
-		Q1 = F1 + Q0;
-		Q2 = 8.0f * (QS - Rsq + QB + F0);
+/*
+ *    calculates pivot matrix such that all the larger elements in the row are on diagonal
+ *
+ *    @param     A,           input matrix matrix
+ *    @param     pivot
+ *    @param     n,           dimenstion of square matrix
+ *    @returns                false = matrix is Singular or non positive definite, true = matrix inversion successful
+ */
+static void mat_pivot(float *A, float *pivot, uint8_t n)
+{
+	for (uint8_t i = 0; i < n; i++) {
+		for (uint8_t j = 0; j < n; j++) {
+			pivot[i * n + j] = (i == j);
+		}
 	}
 
-	*sphere_x = A;
-	*sphere_y = B;
-	*sphere_z = C;
-	*sphere_radius = sqrtf(Rsq);
+	for (uint8_t i = 0; i < n; i++) {
+		uint8_t max_j = i;
 
-	return 0;
+		for (uint8_t j = i; j < n; j++) {
+			if (fabsf(A[j * n + i]) > fabsf(A[max_j * n + i])) {
+				max_j = j;
+			}
+		}
+
+		if (max_j != i) {
+			for (uint8_t k = 0; k < n; k++) {
+				swap(pivot[i * n + k], pivot[max_j * n + k]);
+			}
+		}
+	}
+}
+
+/*
+ *    Decomposes square matrix into Lower and Upper triangular matrices such that
+ *    A*P = L*U, where P is the pivot matrix
+ *    ref: http://rosettacode.org/wiki/LU_decomposition
+ *    @param     U,           upper triangular matrix
+ *    @param     out,         Output inverted upper triangular matrix
+ *    @param     n,           dimension of matrix
+ */
+static void mat_LU_decompose(float *A, float *L, float *U, float *P, uint8_t n)
+{
+	memset(L, 0, n * n * sizeof(float));
+	memset(U, 0, n * n * sizeof(float));
+	memset(P, 0, n * n * sizeof(float));
+	mat_pivot(A, P, n);
+
+	float *APrime = mat_mul(P, A, n);
+
+	for (uint8_t i = 0; i < n; i++) {
+		L[i * n + i] = 1;
+	}
+
+	for (uint8_t i = 0; i < n; i++) {
+		for (uint8_t j = 0; j < n; j++) {
+			if (j <= i) {
+				U[j * n + i] = APrime[j * n + i];
+
+				for (uint8_t k = 0; k < j; k++) {
+					U[j * n + i] -= L[j * n + k] * U[k * n + i];
+				}
+			}
+
+			if (j >= i) {
+				L[j * n + i] = APrime[j * n + i];
+
+				for (uint8_t k = 0; k < i; k++) {
+					L[j * n + i] -= L[j * n + k] * U[k * n + i];
+				}
+
+				L[j * n + i] /= U[i * n + i];
+			}
+		}
+	}
+
+	delete[] APrime;
+}
+
+/*
+ *    calculates matrix inverse of Lower trangular matrix using forward substitution
+ *
+ *    @param     L,           lower triangular matrix
+ *    @param     out,         Output inverted lower triangular matrix
+ *    @param     n,           dimension of matrix
+ */
+static void mat_forward_sub(float *L, float *out, uint8_t n)
+{
+	// Forward substitution solve LY = I
+	for (int i = 0; i < n; i++) {
+		out[i * n + i] = 1 / L[i * n + i];
+
+		for (int j = i + 1; j < n; j++) {
+			for (int k = i; k < j; k++) {
+				out[j * n + i] -= L[j * n + k] * out[k * n + i];
+			}
+
+			out[j * n + i] /= L[j * n + j];
+		}
+	}
+}
+
+/*
+ *    calculates matrix inverse of Upper trangular matrix using backward substitution
+ *
+ *    @param     U,           upper triangular matrix
+ *    @param     out,         Output inverted upper triangular matrix
+ *    @param     n,           dimension of matrix
+ */
+static void mat_back_sub(float *U, float *out, uint8_t n)
+{
+	// Backward Substitution solve UY = I
+	for (int i = n - 1; i >= 0; i--) {
+		out[i * n + i] = 1 / U[i * n + i];
+
+		for (int j = i - 1; j >= 0; j--) {
+			for (int k = i; k > j; k--) {
+				out[j * n + i] -= U[j * n + k] * out[k * n + i];
+			}
+
+			out[j * n + i] /= U[j * n + j];
+		}
+	}
+}
+
+/*
+ *    matrix inverse code for any square matrix using LU decomposition
+ *    inv = inv(U)*inv(L)*P, where L and U are triagular matrices and P the pivot matrix
+ *    ref: http://www.cl.cam.ac.uk/teaching/1314/NumMethods/supporting/mcmaster-kiruba-ludecomp.pdf
+ *    @param     m,           input 4x4 matrix
+ *    @param     inv,      Output inverted 4x4 matrix
+ *    @param     n,           dimension of square matrix
+ *    @returns                false = matrix is Singular, true = matrix inversion successful
+ */
+static bool mat_inverse(float *A, float *inv, uint8_t n)
+{
+	float *L, *U, *P;
+	bool ret = true;
+	L = new float[n * n];
+	U = new float[n * n];
+	P = new float[n * n];
+	mat_LU_decompose(A, L, U, P, n);
+
+	float *L_inv = new float[n * n];
+	float *U_inv = new float[n * n];
+
+	memset(L_inv, 0, n * n * sizeof(float));
+	mat_forward_sub(L, L_inv, n);
+
+	memset(U_inv, 0, n * n * sizeof(float));
+	mat_back_sub(U, U_inv, n);
+
+	// decomposed matrices no longer required
+	delete[] L;
+	delete[] U;
+
+	float *inv_unpivoted = mat_mul(U_inv, L_inv, n);
+	float *inv_pivoted = mat_mul(inv_unpivoted, P, n);
+
+	//check sanity of results
+	for (uint8_t i = 0; i < n; i++) {
+		for (uint8_t j = 0; j < n; j++) {
+			if (!PX4_ISFINITE(inv_pivoted[i * n + j])) {
+				ret = false;
+			}
+		}
+	}
+
+	memcpy(inv, inv_pivoted, n * n * sizeof(float));
+
+	//free memory
+	delete[] inv_pivoted;
+	delete[] inv_unpivoted;
+	delete[] P;
+	delete[] U_inv;
+	delete[] L_inv;
+	return ret;
 }
 
 int ellipsoid_fit_least_squares(const float x[], const float y[], const float z[],
