+++ src/modules/commander/calibration_routines.h
@@ -54,9 +54,6 @@
  *
  * @return 0 on success, 1 on failure
  */
-int sphere_fit_least_squares(const float x[], const float y[], const float z[],
-			     unsigned int size, unsigned int max_iterations, float delta, float *sphere_x, float *sphere_y, float *sphere_z,
-			     float *sphere_radius);
 int ellipsoid_fit_least_squares(const float x[], const float y[], const float z[],
 				unsigned int size, int max_iterations, float *offset_x, float *offset_y, float *offset_z,
 				float *sphere_radius, float *diag_x, float *diag_y, float *diag_z, float *offdiag_x, float *offdiag_y,
@@ -69,8 +66,6 @@ int run_lm_ellipsoid_fit(const float x[], const float y[], const float z[], floa
 			 unsigned int size, float *offset_x, float *offset_y, float *offset_z,
 			 float *sphere_radius, float *diag_x, float *diag_y, float *diag_z, float *offdiag_x, float *offdiag_y,
 			 float *offdiag_z);
-bool inverse4x4(float m[], float invOut[]);
-bool mat_inverse(float *A, float *inv, uint8_t n);
 
 // The order of these cannot change since the calibration calculations depend on them in this order
 enum detect_orientation_return {
@@ -82,7 +77,7 @@ enum detect_orientation_return {
 	DETECT_ORIENTATION_RIGHTSIDE_UP,
 	DETECT_ORIENTATION_ERROR
 };
-static const unsigned detect_orientation_side_count = 6;
+static constexpr unsigned detect_orientation_side_count = 6;
 
 /// Wait for vehicle to become still and detect it's orientation
 ///	@return Returns detect_orientation_return according to orientation when vehicle
